+++ RougeLike/Source/Utilities/Shapes/Circle.cpp.cpp
@@ -5,7 +5,7 @@ Circle::Circle(VectorF center, float radius, int pointCount)
 {
 	ASSERT(Warning, pointCount > 5, "A circle needs at least 6 points to be a circle, this one has %d\n", pointCount);
 
-	float arc = 360 / pointCount;
+	float arc = (float)(360 / pointCount);
 
 	// pointing up in the (0, -1) direction
 	VectorF radiusLine(0.0f, -radius);
